refactor: changed references to calendar into folder
