+++ src/integrations/shared-invite-reply/index.tsx
@@ -208,7 +208,7 @@ const SharedCalendarResponse: FC<SharedCalendarResponseReturnType> = ({
 							view={view}
 							rid={rid}
 							msgId={mailMsg.id}
-							sharedCalendarName={sharedContentItem}
+							sharedFolderName={sharedContentItem}
 							grantee={grantee}
 							owner={owner}
 							role={role}
